fix remaining build error on windows
